Refactor visit() as TypeScript
