+++ rollup.dev-config.js
@@ -6,8 +6,7 @@ export default {
     'src/ast/index.js',
     'src/doc/Document.js',
     'src/errors.js',
-    'src/options.js',
-    'src/visit.js'
+    'src/options.js'
   ],
   external: [resolve('src/doc/directives.js')],
   output: { dir: 'lib', format: 'cjs', esModule: false, preserveModules: true },
